+++ src/aux-common/runtime/AuxLibrary.spec.ts
@@ -4488,6 +4488,26 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     false,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    'myEndpoint'
+                );
+                const expected = recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    false,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4510,6 +4530,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     false,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4534,6 +4555,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     false,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4553,6 +4575,26 @@ describe('AuxLibrary', () => {
                     'address',
                     { data: true },
                     true,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.recordManualApprovalData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    'myEndpoint',
+                );
+                const expected = recordData(
+                    'recordKey',
+                    'address',
+                    { data: true },
+                    true,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4575,6 +4617,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     true,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4599,6 +4642,7 @@ describe('AuxLibrary', () => {
                         },
                     },
                     true,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4616,6 +4660,24 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.getData(
+                    'recordKey',
+                    'address',
+                    'myEndpoint',
+                );
+                const expected = getRecordData(
+                    'recordKey',
+                    'address',
+                    false,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4631,6 +4693,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     false,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4648,6 +4711,24 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.getManualApprovalData(
+                    'recordKey',
+                    'address',
+                    'myEndpoint'
+                );
+                const expected = getRecordData(
+                    'recordKey',
+                    'address',
+                    true,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4663,6 +4744,7 @@ describe('AuxLibrary', () => {
                     'recordName',
                     'address',
                     true,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4676,6 +4758,19 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     null,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.listData('recordName', undefined, 'myEndpoint');
+                const expected = listDataRecord(
+                    'recordName',
+                    null,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4690,6 +4785,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     'address',
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4704,6 +4800,7 @@ describe('AuxLibrary', () => {
                 const expected = listDataRecord(
                     'recordName',
                     'address',
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4721,6 +4818,24 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     false,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.eraseData(
+                    'recordKey',
+                    'address',
+                    'myEndpoint',
+                );
+                const expected = eraseRecordData(
+                    'recordKey',
+                    'address',
+                    false,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4762,6 +4877,24 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'address',
                     true,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.eraseManualApprovalData(
+                    'recordKey',
+                    'address',
+                    'myEndpoint'
+                );
+                const expected = eraseRecordData(
+                    'recordKey',
+                    'address',
+                    true,
+                    'myEndpoint',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4810,6 +4943,28 @@ describe('AuxLibrary', () => {
                     'data',
                     'description',
                     undefined,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.recordFile(
+                    'recordKey',
+                    'data',
+                    {
+                        description: 'description',
+                    },
+                    'https://localhost:5000'
+                );
+                const expected = recordFile(
+                    'recordKey',
+                    'data',
+                    'description',
+                    undefined,
+                    'https://localhost:5000',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4826,6 +4981,7 @@ describe('AuxLibrary', () => {
                     'data',
                     undefined,
                     undefined,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4845,6 +5001,7 @@ describe('AuxLibrary', () => {
                     'data',
                     undefined,
                     'image/png',
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4885,6 +5042,7 @@ describe('AuxLibrary', () => {
                     },
                     undefined,
                     undefined,
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4961,6 +5119,23 @@ describe('AuxLibrary', () => {
                 const expected = eraseFile(
                     'recordKey',
                     'fileUrl',
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.eraseFile(
+                    'recordKey',
+                    'fileUrl',
+                    'http://localhost:5000'
+                );
+                const expected = eraseFile(
+                    'recordKey',
+                    'fileUrl',
+                    'http://localhost:5000',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -4998,6 +5173,24 @@ describe('AuxLibrary', () => {
                     'recordKey',
                     'eventName',
                     1,
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.recordEvent(
+                    'recordKey',
+                    'eventName',
+                    'http://localhost:5000'
+                );
+                const expected = recordEvent(
+                    'recordKey',
+                    'eventName',
+                    1,
+                    'http://localhost:5000',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5038,6 +5231,23 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordKey',
                     'eventName',
+                    null,
+                    context.tasks.size
+                );
+                expect(action[ORIGINAL_OBJECT]).toEqual(expected);
+                expect(context.actions).toEqual([expected]);
+            });
+
+            it('should support custom endpoints', async () => {
+                const action: any = library.api.os.countEvents(
+                    'recordKey',
+                    'eventName',
+                    'http://localhost:5000'
+                );
+                const expected = getEventCount(
+                    'recordKey',
+                    'eventName',
+                    'http://localhost:5000',
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);
@@ -5052,6 +5262,7 @@ describe('AuxLibrary', () => {
                 const expected = getEventCount(
                     'recordName',
                     'eventName',
+                    null,
                     context.tasks.size
                 );
                 expect(action[ORIGINAL_OBJECT]).toEqual(expected);